♻️ refactor: home 컴포넌트 분리 및 useSwipeableCard 훅 추출
