+++ src/components/home/MealTimetableWidget.tsx
@@ -1,16 +1,5 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import { useState, useEffect, useRef, useCallback } from 'react';
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  Animated,
-  Easing,
-  PanResponder,
-  ScrollView,
-  ActivityIndicator,
-  useWindowDimensions,
-} from 'react-native';
+import { useState, useEffect } from 'react';
+import { View, Text, TouchableOpacity, useWindowDimensions } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { router } from 'expo-router';
 import { useTranslation } from 'react-i18next';
@@ -20,376 +9,21 @@ import {
   getMealsForChild,
   fetchSchoolTimetables,
   getPeriodsForChild,
-  MealMenu,
-  TimetablePeriod,
 } from '../../api/meal';
-import { ChildItem } from '../../api/child';
+import type { MealMenu, TimetablePeriod } from '../../api/meal';
 import colors from '../../constants/colors';
-import styles from '../../styles/home/mealTimetableWidget';
+import styles from './mealTimetableWidget.styles';
+import MealCard from './MealCard';
+import TimetableCard from './TimetableCard';
 
 const CARD_GAP = 16;
 const HORIZONTAL_PADDING = 40;
-const SLIDE_DURATION = 550;
-const AUTO_INTERVAL = 5000;
 
 const getTodayStr = (): string => {
   const d = new Date();
   const m = String(d.getMonth() + 1).padStart(2, '0');
   const day = String(d.getDate()).padStart(2, '0');
   return `${d.getFullYear()}-${m}-${day}`;
-};
-
-interface MealCardProps {
-  cardWidth: number;
-  items: ChildItem[];
-  meals: MealMenu[][];
-  loading: boolean;
-  fontScale: number;
-}
-
-const MealCard = ({ cardWidth, items, meals, loading, fontScale }: MealCardProps) => {
-  const { t } = useTranslation();
-  const [activeIdx, setActiveIdx] = useState(0);
-  const activeIdxRef = useRef(0);
-  const isPausedRef = useRef(false);
-  const translateX = useRef(new Animated.Value(0)).current;
-  const activeChild = items[activeIdx];
-
-  const slideTo = useCallback(
-    (idx: number) => {
-      activeIdxRef.current = idx;
-      setActiveIdx(idx);
-      Animated.timing(translateX, {
-        toValue: -idx * cardWidth,
-        duration: SLIDE_DURATION,
-        easing: Easing.out(Easing.cubic),
-        useNativeDriver: true,
-      }).start();
-    },
-    [cardWidth, translateX]
-  );
-
-  const slideToRef = useRef(slideTo);
-  slideToRef.current = slideTo;
-  const itemsLengthRef = useRef(items.length);
-  itemsLengthRef.current = items.length;
-
-  useEffect(() => {
-    if (!items.length) return;
-    const clamped = Math.min(activeIdxRef.current, items.length - 1);
-    if (clamped !== activeIdxRef.current) {
-      activeIdxRef.current = clamped;
-      setActiveIdx(clamped);
-    }
-    translateX.setValue(-clamped * cardWidth);
-  }, [items.length, cardWidth, translateX]);
-
-  useEffect(() => {
-    if (items.length <= 1) return undefined;
-    const interval = setInterval(() => {
-      if (!isPausedRef.current) {
-        const next = (activeIdxRef.current + 1) % items.length;
-        slideTo(next);
-      }
-    }, AUTO_INTERVAL);
-    return () => clearInterval(interval);
-  }, [items.length, slideTo]);
-
-  const panResponder = useRef(
-    PanResponder.create({
-      onMoveShouldSetPanResponder: (_, g) => Math.abs(g.dx) > 8 && Math.abs(g.dx) > Math.abs(g.dy),
-      onPanResponderRelease: (_, g) => {
-        const len = itemsLengthRef.current;
-        if (g.dx < -40) {
-          slideToRef.current((activeIdxRef.current + 1) % len);
-        } else if (g.dx > 40) {
-          slideToRef.current((activeIdxRef.current - 1 + len) % len);
-        }
-      },
-    })
-  ).current;
-
-  const renderMenuContent = (menus: MealMenu[]) => {
-    if (loading) {
-      return <ActivityIndicator color={colors.primary[400]} style={styles.loadingIndicator} />;
-    }
-    if (!menus.length) {
-      return <Text style={styles.emptyText}>{t('mealTimetable.meal.empty')}</Text>;
-    }
-    return (
-      <View style={styles.menuList}>
-        {menus.map((menu) => (
-          <View key={menu.name} style={styles.menuItem}>
-            <View style={styles.menuBullet} />
-            <Text style={styles.menuName} numberOfLines={2}>
-              {menu.name}
-              {menu.allergyNums.length > 0 && (
-                <Text style={styles.allergyText}> ({menu.allergyNums.join('.')})</Text>
-              )}
-            </Text>
-          </View>
-        ))}
-      </View>
-    );
-  };
-
-  return (
-    <View
-      style={styles.card}
-      onTouchStart={() => {
-        isPausedRef.current = true;
-      }}
-      onTouchEnd={() => {
-        isPausedRef.current = false;
-      }}
-      onTouchCancel={() => {
-        isPausedRef.current = false;
-      }}
-    >
-      <View style={styles.cardFixedSection}>
-        <View style={styles.cardHeader}>
-          <View style={[styles.iconBox, { backgroundColor: colors.primary[100] }]}>
-            <Ionicons name="restaurant-outline" size={22} color={colors.primary[500]} />
-          </View>
-          <View style={styles.headerTexts}>
-            {fontScale < 1.3 && (
-              <Text style={[styles.cardLabel, { color: colors.primary[500] }]}>
-                {t('mealTimetable.meal.cardLabel')}
-              </Text>
-            )}
-            <Text style={styles.cardTitle} numberOfLines={1}>
-              {t('mealTimetable.meal.cardTitle')}
-            </Text>
-          </View>
-        </View>
-        <View style={styles.childRow}>
-          <View style={[styles.childDot, { backgroundColor: activeChild.colorCode }]} />
-          <Text style={styles.childText} numberOfLines={1}>
-            {activeChild.name} · {activeChild.schoolName}
-          </Text>
-        </View>
-        <View style={styles.divider} />
-      </View>
-
-      <View
-        style={styles.menuAreaContainer}
-        {...(items.length > 1 ? panResponder.panHandlers : {})}
-      >
-        <Animated.View
-          style={[
-            styles.slideRow,
-            { width: cardWidth * items.length, transform: [{ translateX }] },
-          ]}
-        >
-          {items.map((item, index) => (
-            <View key={String(item.id)} style={[{ width: cardWidth }, styles.menuPageContent]}>
-              <ScrollView
-                showsVerticalScrollIndicator={false}
-                nestedScrollEnabled
-                style={styles.scrollArea}
-              >
-                {renderMenuContent(meals[index] ?? [])}
-              </ScrollView>
-            </View>
-          ))}
-        </Animated.View>
-      </View>
-
-      {items.length > 1 && (
-        <View style={styles.dotsRow}>
-          {items.map((item, idx) => (
-            <View
-              key={item.id}
-              style={[
-                styles.dot,
-                { backgroundColor: idx === activeIdx ? colors.primary[500] : colors.gray[200] },
-              ]}
-            />
-          ))}
-        </View>
-      )}
-    </View>
-  );
-};
-
-interface TimetableCardProps {
-  cardWidth: number;
-  items: ChildItem[];
-  timetables: TimetablePeriod[][];
-  loading: boolean;
-  fontScale: number;
-}
-
-const TimetableCard = ({
-  cardWidth,
-  items,
-  timetables,
-  loading,
-  fontScale,
-}: TimetableCardProps) => {
-  const { t } = useTranslation();
-  const [activeIdx, setActiveIdx] = useState(0);
-  const activeIdxRef = useRef(0);
-  const isPausedRef = useRef(false);
-  const translateX = useRef(new Animated.Value(0)).current;
-  const activeChild = items[activeIdx];
-
-  const slideTo = useCallback(
-    (idx: number) => {
-      activeIdxRef.current = idx;
-      setActiveIdx(idx);
-      Animated.timing(translateX, {
-        toValue: -idx * cardWidth,
-        duration: SLIDE_DURATION,
-        easing: Easing.out(Easing.cubic),
-        useNativeDriver: true,
-      }).start();
-    },
-    [cardWidth, translateX]
-  );
-
-  const slideToRef = useRef(slideTo);
-  slideToRef.current = slideTo;
-  const itemsLengthRef = useRef(items.length);
-  itemsLengthRef.current = items.length;
-
-  useEffect(() => {
-    if (!items.length) return;
-    const clamped = Math.min(activeIdxRef.current, items.length - 1);
-    if (clamped !== activeIdxRef.current) {
-      activeIdxRef.current = clamped;
-      setActiveIdx(clamped);
-    }
-    translateX.setValue(-clamped * cardWidth);
-  }, [items.length, cardWidth, translateX]);
-
-  useEffect(() => {
-    if (items.length <= 1) return undefined;
-    const interval = setInterval(() => {
-      if (!isPausedRef.current) {
-        const next = (activeIdxRef.current + 1) % items.length;
-        slideTo(next);
-      }
-    }, AUTO_INTERVAL);
-    return () => clearInterval(interval);
-  }, [items.length, slideTo]);
-
-  const panResponder = useRef(
-    PanResponder.create({
-      onMoveShouldSetPanResponder: (_, g) => Math.abs(g.dx) > 8 && Math.abs(g.dx) > Math.abs(g.dy),
-      onPanResponderRelease: (_, g) => {
-        const len = itemsLengthRef.current;
-        if (g.dx < -40) {
-          slideToRef.current((activeIdxRef.current + 1) % len);
-        } else if (g.dx > 40) {
-          slideToRef.current((activeIdxRef.current - 1 + len) % len);
-        }
-      },
-    })
-  ).current;
-
-  const renderTimetableContent = (periods: TimetablePeriod[]) => {
-    if (loading) {
-      return <ActivityIndicator color={colors.secondary[500]} style={styles.loadingIndicator} />;
-    }
-    if (!periods.length) {
-      return <Text style={styles.emptyText}>{t('mealTimetable.timetable.empty')}</Text>;
-    }
-    return (
-      <View style={styles.periodList}>
-        {periods.map((p) => (
-          <View key={p.period} style={styles.periodItem}>
-            <View style={styles.periodBadge}>
-              <Text style={styles.periodBadgeText}>
-                {t('mealTimetable.timetable.period', { period: p.period })}
-              </Text>
-            </View>
-            <Text style={styles.periodSubject} numberOfLines={1}>
-              {p.subject}
-            </Text>
-          </View>
-        ))}
-      </View>
-    );
-  };
-
-  return (
-    <View
-      style={styles.card}
-      onTouchStart={() => {
-        isPausedRef.current = true;
-      }}
-      onTouchEnd={() => {
-        isPausedRef.current = false;
-      }}
-      onTouchCancel={() => {
-        isPausedRef.current = false;
-      }}
-    >
-      <View style={styles.cardFixedSection}>
-        <View style={styles.cardHeader}>
-          <View style={[styles.iconBox, { backgroundColor: colors.secondary[100] }]}>
-            <Ionicons name="alarm-outline" size={22} color={colors.secondary[600]} />
-          </View>
-          <View style={styles.headerTexts}>
-            {fontScale < 1.3 && (
-              <Text style={[styles.cardLabel, { color: colors.secondary[600] }]}>
-                {t('mealTimetable.timetable.cardLabel')}
-              </Text>
-            )}
-            <Text style={styles.cardTitle} numberOfLines={1}>
-              {t('mealTimetable.timetable.cardTitle')}
-            </Text>
-          </View>
-        </View>
-        <View style={styles.childRow}>
-          <View style={[styles.childDot, { backgroundColor: activeChild.colorCode }]} />
-          <Text style={styles.childText} numberOfLines={1}>
-            {activeChild.name} · {activeChild.schoolName}
-          </Text>
-        </View>
-        <View style={styles.divider} />
-      </View>
-
-      <View
-        style={styles.menuAreaContainer}
-        {...(items.length > 1 ? panResponder.panHandlers : {})}
-      >
-        <Animated.View
-          style={[
-            styles.slideRow,
-            { width: cardWidth * items.length, transform: [{ translateX }] },
-          ]}
-        >
-          {items.map((item, index) => (
-            <View key={String(item.id)} style={[{ width: cardWidth }, styles.menuPageContent]}>
-              <ScrollView
-                showsVerticalScrollIndicator={false}
-                nestedScrollEnabled
-                style={styles.scrollArea}
-              >
-                {renderTimetableContent(timetables[index] ?? [])}
-              </ScrollView>
-            </View>
-          ))}
-        </Animated.View>
-      </View>
-
-      {items.length > 1 && (
-        <View style={styles.dotsRow}>
-          {items.map((item, idx) => (
-            <View
-              key={item.id}
-              style={[
-                styles.dot,
-                { backgroundColor: idx === activeIdx ? colors.secondary[500] : colors.gray[200] },
-              ]}
-            />
-          ))}
-        </View>
-      )}
-    </View>
-  );
 };
 
 const MealTimetableWidget = () => {
